fix chronometr Conversatiom fragment
