+++ sample-videochat-webrtc/src/main/java/com/quickblox/sample/groupchatwebrtc/activities/CallActivity.java
@@ -463,8 +463,6 @@ public class CallActivity extends BaseActivity implements QBRTCClientSessionCall
                 if (isInCommingCall) {
                     stopIncomeCallTimer();
                 }
-
-//                startTimer();
                 Log.d(TAG, "onConnectedToUser() is started");
 
             }
